Australian Bureau of Statistics (ABS) map data

These data sets are simplifications of the formal statistical areas and regions published by the ABS in 2016.

abs_ste

Format

Simple features data frame with

NAME

Area name

geometry

Geometry column in 'sfc' format

Details

Each layer was read from the source file with 'sf' package and attributes were removed, leaving a single 'NAME' column from the year-specific column names. The geometry has been simplified using `rmapshaper::ms_simplify` with default arguments (0.05 detail).

Several layers are not included from the total available.

The entire nation layer 'AUST' is not included as it is the union of the State and Territory layer.

Statistical Areas Level 1 is not included as it is very large (56Mb after simplification).

The mesh blocks are not included.

Data layers

abs_ced

Commonwealth Electoral Divisions

abs_gccsa

Greater Capital City Statistical Areas

abs_ireg

Indigenous Regions

abs_lga

Local Goverment Areas

abs_ra

Remoteness Structure

abs_sa2

Statistical Areas Level 2

abs_sa3

Statistical Areas Level 3

abs_sa4

Statistical Areas Level 4

abs_sed

State Electoral Divisions

abs_ste

State and Territory

See also